fix: merge engine and editor `Editable` components
The `Editable` component was split across two files: an outer
`PortableTextEditable` in `editor/Editable.tsx` wrapping an inner
`EngineEditable` in `engine/react/components/editable.tsx`. The seam was
bogus: the inner already imports `EditorActor` from `editor/editor-machine`,
dom-traversal and node-traversal utilities, and sends behavior events
directly. Every DOM event ran through both halves via `isEventHandled`
short-circuits glued on `attributes.onX` props.

This merges the two into a single `PortableTextEditable`. The render-prop
types (`RenderElementProps`, `RenderLeafProps`, `RenderTextProps`) move to
`editor/render-props-types.ts` so engine-internal components can reach
them without crossing back through the merged file. The event-handler
ladder collapses: each native handler now runs the consumer hook (or
behavior-event raise) first, bails on `defaultPrevented` or
`isPropagationStopped`, then runs the engine's DOM logic.

The file stays exempt from react-compiler. The engine has long-lived
mutable instance fields (`editor.composing`, `editor.focused`,
`editor.userSelection`, `editor.domElement`) that this file assigns to as
part of translating raw DOM events into engine state; those mutations are
the bridge's job, not accidental React. Previously the exemption rode on
the wholesale `src/engine/**` ignore; now it lives as a per-file rule in
`eslint.config.js` and `package.config.ts`.
