+++ .changeset/merge-editable-components.md
@@ -1,0 +1,5 @@
+---
+'@portabletext/editor': patch
+---
+
+fix: merge engine and editor `Editable` components
