+++ packages/editor/src/editor/render.element.tsx
@@ -7,7 +7,6 @@ import {useSelector} from '@xstate/react'
 import {useContext, useMemo, type ReactElement} from 'react'
 import type {DropPosition} from '../behaviors/behavior.core.drop-position'
 import type {Path} from '../engine/interfaces/path'
-import type {RenderElementProps} from '../engine/react/components/editable'
 import {useEngineStatic} from '../engine/react/hooks/use-engine-static'
 import {isInline} from '../node-traversal/is-inline'
 import {serializePath} from '../paths/serialize-path'
@@ -27,6 +26,7 @@ import type {LegacyRenderHooks} from './legacy-render-hooks'
 import {NewPipelineContext} from './new-pipeline-context'
 import {ParentContainerContext} from './parent-container-context'
 import {ParentTextBlockContext} from './parent-text-block-context'
+import type {RenderElementProps} from './render-props-types'
 import {RenderBlockObject} from './render.block-object'
 import {RenderContainer} from './render.container'
 import {renderDefaultTextBlock} from './render.default'
