+++ packages/editor/src/engine/react/components/object-node.tsx
@@ -1,13 +1,13 @@
 import type {PortableTextObject} from '@portabletext/schema'
 import React, {useContext, type JSX} from 'react'
 import {NewPipelineContext} from '../../../editor/new-pipeline-context'
+import type {RenderElementProps} from '../../../editor/render-props-types'
 import {serializePath} from '../../../paths/serialize-path'
 import {isElementDecorationsEqual} from '../../dom/utils/range-list'
 import type {Path} from '../../interfaces/path'
 import type {DecoratedRange} from '../../interfaces/text'
 import {pathEquals} from '../../path/path-equals'
 import {useReadOnly} from '../hooks/use-read-only'
-import type {RenderElementProps} from './editable'
 
 /**
  * Wrapper for block or inline object nodes that have no children in the engine
